+++ app/controllers/user.go
@@ -4,90 +4,15 @@ import (
 	"github.com/revel/revel"
 	"fmt"
 	"strconv"
-	"github.com/gocql/gocql"
+	. "MyRevelApp/app/models"
+	"MyRevelApp/app"
 )
 
-type PhoneNum struct {
-	Id int
-	ContactId gocql.UUID
-	Phonenumber string
-
-}
-
-type Contact struct{
-	Id gocql.UUID
-	FirstName string
-	LastName string
-	Email string
-	PhoneNumbersStamped []PhoneNum
-	PhoneNumbers []string
-
-}
-
-type UserContancts struct {
-	UserName string
-	Password string
-	Contacts []Contact
-
-}
 
 type User struct {
 	*revel.Controller
-	Db *gocql.Session
 }
 
-func (contact  *Contact)StampContactId () {
-	i :=0
-	contact.PhoneNumbersStamped = []PhoneNum{}
-	fmt.Println(len(contact.PhoneNumbers))
-	contactid := contact.Id
-	for i<len(contact.PhoneNumbers){
-		numberid := i
-		phonenumber := contact.PhoneNumbers[i]
-		contact.PhoneNumbersStamped = append(contact.PhoneNumbersStamped , PhoneNum{ContactId:contactid , Id:numberid , Phonenumber:phonenumber})
-		i++
-	}
-}
-
-func (user * UserContancts) DeleteContact(id string , db *gocql.Session) error{
-	err := db.Query("delete from user_data where username = ? and contact_id = ?",user.UserName , id).Exec()
-	fmt.Println(err)
-	return err
-
-}
-///////////////////////////////////////////////////////////////////////////////////////////////////////
-func (user * UserContancts) DeleteContactNumber(id string , contactid string , db *gocql.Session) error{
-	err := db.Query("delete contact_phonenumbers[?] from user_data where username = ? and contact_id = ?",id ,user.UserName , contactid ).Exec()
-	return err
-
-}
-///////////////////////////////////////////////////////////////////////////////////////////////////////
-func (user * UserContancts) GetUserContacts( db *gocql.Session) error{
-	var newcontact Contact
-	rows := db.Query("select contact_id,contact_email,contact_fname,contact_lname,contact_phonenumbers from user_data where username= ?" , user.UserName)
-	scanner :=rows.Iter().Scanner()
-	for scanner.Next(){
-		scanner.Scan(&newcontact.Id , &newcontact.Email, &newcontact.FirstName , &newcontact.LastName , &newcontact.PhoneNumbers)
-		newcontact.StampContactId()
-
-		user.Contacts = append(user.Contacts, newcontact)
-	}
-	err := rows.Iter().Close()
-	return err
-}
-///////////////////////////////////////////////////////////////////////////////////////////////////////
-func (user * UserContancts) InsertNewContact(c Contact , db * gocql.Session) (Contact, error){
-
-	err := db.Query("insert into user_data (username ,contact_id , contact_email , contact_fname , contact_lname , contact_phonenumbers ) values(? , uuid() , ? , ? , ? , ? ) ", user.UserName, c.Email , c.FirstName, c.LastName, c.PhoneNumbers).Exec()
-	if err !=nil {
-		fmt.Println(err)
-		return Contact{} , err
-	}
-	c.StampContactId()
-	user.Contacts = append(user.Contacts, c)
-	return c , nil
-
-}
 ///////////////////////////////////////////////////////////////////////////////////////////////////////
 func (user *User) Userpage() revel.Result {
 
@@ -101,7 +26,7 @@ func (user *User) Userpage() revel.Result {
 	myuser.UserName = Username
 
 
-	err :=myuser.GetUserContacts(user.Db)
+	err :=myuser.GetUserContacts(app.DB)
 	if err != nil {
 		user.RenderError(err)
 
@@ -112,6 +37,7 @@ func (user *User) Userpage() revel.Result {
 func (user *User) AddContact() revel.Result{
 	myuser := UserContancts{}
 	myuser.UserName=user.Session["user"]
+
 	user.Validation.Required(user.Params.Get("first-name"))
 	user.Validation.Required(user.Params.Get("last-name"))
 	user.Validation.Required(user.Params.Get("email"))
@@ -134,17 +60,17 @@ func (user *User) AddContact() revel.Result{
 			phonenumbers = append(phonenumbers,str)
 			i++
 		}
-		fmt.Println(phonenumbers)
 		c := Contact{
 			FirstName:user.Params.Get("first-name"),
 			LastName:user.Params.Get("last-name"),
 			Email:user.Params.Get("email"),
 			PhoneNumbers:phonenumbers,
 		}
-		c , err := myuser.InsertNewContact(c , user.Db)
+		c , err := myuser.InsertNewContact(c , app.DB)
 		if err !=nil {
 			return user.RenderError(err)
 		}
+		fmt.Println(c)
 		return user.RenderJSON(c)
 	}
 }
@@ -156,7 +82,7 @@ func (user User) Delete() revel.Result {
 		return user.RenderTemplate("user/userpage.html")
 	} else {
 		myuser :=UserContancts{UserName:user.Session["user"]}
-		err := myuser.DeleteContact(user.Params.Get("id") , user.Db)
+		err := myuser.DeleteContact(user.Params.Get("id") , app.DB)
 		if err != nil {
 			return user.RenderError(err)
 		}
@@ -171,7 +97,7 @@ func (user User) DeleteNum() revel.Result{
 		return user.RenderTemplate("user/userpage.html")
 	} else {
 		myuser :=UserContancts{UserName:user.Session["user"]}
-		err := myuser.DeleteContactNumber(user.Params.Get("id") , user.Params.Get("ID") , user.Db)
+		err := myuser.DeleteContactNumber(user.Params.Get("id") , user.Params.Get("ID") , app.DB)
 		if err != nil {
 			return user.RenderError(err)
 		}
@@ -181,22 +107,6 @@ func (user User) DeleteNum() revel.Result{
 ///////////////////////////////////////////////////////////////////////////////////////////////////////
 func (user User) Logout() revel.Result{
 	user.Session["user"] = ""
-
-
 	return user.Redirect("/")
 }
 ///////////////////////////////////////////////////////////////////////////////////////////////////////
-func startDatabase(user *User) revel.Result{
-	var err error
-	cluster := gocql.NewCluster("127.0.0.1")
-	cluster.Keyspace = "address_book"
-	user.Db, err= cluster.CreateSession()
-	if err != nil {
-		panic(err)
-	}
-	return nil
-}
-///////////////////////////////////////////////////////////////////////////////////////////////////////
-func init(){
-	revel.InterceptMethod(startDatabase , revel.BEFORE)
-}
