+++ app/routes/routes.go
@@ -16,20 +16,16 @@ func (_ tApp) Index(
 }
 
 func (_ tApp) SignIn(
-		logins controllers.Logins,
 		) string {
 	args := make(map[string]string)
 	
-	revel.Unbind(args, "logins", logins)
 	return revel.MainRouter.Reverse("App.SignIn", args).URL
 }
 
 func (_ tApp) Register(
-		logins controllers.Logins,
 		) string {
 	args := make(map[string]string)
 	
-	revel.Unbind(args, "logins", logins)
 	return revel.MainRouter.Reverse("App.Register", args).URL
 }
 
@@ -81,6 +77,35 @@ func (_ tUser) Logout(
 }
 
 
+type tStatic struct {}
+var Static tStatic
+
+
+func (_ tStatic) Serve(
+		prefix string,
+		filepath string,
+		) string {
+	args := make(map[string]string)
+	
+	revel.Unbind(args, "prefix", prefix)
+	revel.Unbind(args, "filepath", filepath)
+	return revel.MainRouter.Reverse("Static.Serve", args).URL
+}
+
+func (_ tStatic) ServeModule(
+		moduleName string,
+		prefix string,
+		filepath string,
+		) string {
+	args := make(map[string]string)
+	
+	revel.Unbind(args, "moduleName", moduleName)
+	revel.Unbind(args, "prefix", prefix)
+	revel.Unbind(args, "filepath", filepath)
+	return revel.MainRouter.Reverse("Static.ServeModule", args).URL
+}
+
+
 type tTestRunner struct {}
 var TestRunner tTestRunner
 
@@ -117,35 +142,6 @@ func (_ tTestRunner) List(
 	args := make(map[string]string)
 	
 	return revel.MainRouter.Reverse("TestRunner.List", args).URL
-}
-
-
-type tStatic struct {}
-var Static tStatic
-
-
-func (_ tStatic) Serve(
-		prefix string,
-		filepath string,
-		) string {
-	args := make(map[string]string)
-	
-	revel.Unbind(args, "prefix", prefix)
-	revel.Unbind(args, "filepath", filepath)
-	return revel.MainRouter.Reverse("Static.Serve", args).URL
-}
-
-func (_ tStatic) ServeModule(
-		moduleName string,
-		prefix string,
-		filepath string,
-		) string {
-	args := make(map[string]string)
-	
-	revel.Unbind(args, "moduleName", moduleName)
-	revel.Unbind(args, "prefix", prefix)
-	revel.Unbind(args, "filepath", filepath)
-	return revel.MainRouter.Reverse("Static.ServeModule", args).URL
 }
