+++ app/tmp/main.go
@@ -8,9 +8,9 @@ import (
 	_ "MyRevelApp/app"
 	controllers "MyRevelApp/app/controllers"
 	tests "MyRevelApp/tests"
-	controllers1 "github.com/revel/modules/static/app/controllers"
+	controllers0 "github.com/revel/modules/static/app/controllers"
 	_ "github.com/revel/modules/testrunner/app"
-	controllers0 "github.com/revel/modules/testrunner/app/controllers"
+	controllers1 "github.com/revel/modules/testrunner/app/controllers"
 	"github.com/revel/revel/testing"
 )
 
@@ -36,14 +36,13 @@ func main() {
 				Args: []*revel.MethodArg{ 
 				},
 				RenderArgNames: map[int][]string{ 
-					25: []string{ 
+					22: []string{ 
 					},
 				},
 			},
 			&revel.MethodType{
 				Name: "SignIn",
 				Args: []*revel.MethodArg{ 
-					&revel.MethodArg{Name: "logins", Type: reflect.TypeOf((*controllers.Logins)(nil)) },
 				},
 				RenderArgNames: map[int][]string{ 
 				},
@@ -51,7 +50,6 @@ func main() {
 			&revel.MethodType{
 				Name: "Register",
 				Args: []*revel.MethodArg{ 
-					&revel.MethodArg{Name: "logins", Type: reflect.TypeOf((*controllers.Logins)(nil)) },
 				},
 				RenderArgNames: map[int][]string{ 
 				},
@@ -61,7 +59,7 @@ func main() {
 				Args: []*revel.MethodArg{ 
 				},
 				RenderArgNames: map[int][]string{ 
-					136: []string{ 
+					109: []string{ 
 					},
 				},
 			},
@@ -75,7 +73,7 @@ func main() {
 				Args: []*revel.MethodArg{ 
 				},
 				RenderArgNames: map[int][]string{ 
-					109: []string{ 
+					34: []string{ 
 						"myuser",
 					},
 				},
@@ -111,7 +109,31 @@ func main() {
 			
 		})
 	
-	revel.RegisterController((*controllers0.TestRunner)(nil),
+	revel.RegisterController((*controllers0.Static)(nil),
+		[]*revel.MethodType{
+			&revel.MethodType{
+				Name: "Serve",
+				Args: []*revel.MethodArg{ 
+					&revel.MethodArg{Name: "prefix", Type: reflect.TypeOf((*string)(nil)) },
+					&revel.MethodArg{Name: "filepath", Type: reflect.TypeOf((*string)(nil)) },
+				},
+				RenderArgNames: map[int][]string{ 
+				},
+			},
+			&revel.MethodType{
+				Name: "ServeModule",
+				Args: []*revel.MethodArg{ 
+					&revel.MethodArg{Name: "moduleName", Type: reflect.TypeOf((*string)(nil)) },
+					&revel.MethodArg{Name: "prefix", Type: reflect.TypeOf((*string)(nil)) },
+					&revel.MethodArg{Name: "filepath", Type: reflect.TypeOf((*string)(nil)) },
+				},
+				RenderArgNames: map[int][]string{ 
+				},
+			},
+			
+		})
+	
+	revel.RegisterController((*controllers1.TestRunner)(nil),
 		[]*revel.MethodType{
 			&revel.MethodType{
 				Name: "Index",
@@ -152,36 +174,12 @@ func main() {
 			
 		})
 	
-	revel.RegisterController((*controllers1.Static)(nil),
-		[]*revel.MethodType{
-			&revel.MethodType{
-				Name: "Serve",
-				Args: []*revel.MethodArg{ 
-					&revel.MethodArg{Name: "prefix", Type: reflect.TypeOf((*string)(nil)) },
-					&revel.MethodArg{Name: "filepath", Type: reflect.TypeOf((*string)(nil)) },
-				},
-				RenderArgNames: map[int][]string{ 
-				},
-			},
-			&revel.MethodType{
-				Name: "ServeModule",
-				Args: []*revel.MethodArg{ 
-					&revel.MethodArg{Name: "moduleName", Type: reflect.TypeOf((*string)(nil)) },
-					&revel.MethodArg{Name: "prefix", Type: reflect.TypeOf((*string)(nil)) },
-					&revel.MethodArg{Name: "filepath", Type: reflect.TypeOf((*string)(nil)) },
-				},
-				RenderArgNames: map[int][]string{ 
-				},
-			},
-			
-		})
-	
 	revel.DefaultValidationKeys = map[string]map[int]string{ 
 		"MyRevelApp/app/controllers.App.Login": { 
-			114: "username",
-			115: "password",
-			116: "username",
-			117: "password",
+			91: "logins.Username",
+			92: "logins.Password",
+			93: "logins.Username",
+			94: "logins.Password",
 		},
 	}
 	testing.TestSuites = []interface{}{
