+++ app/views/User/userpage.html
@@ -250,7 +250,7 @@ function AddContact()
             type: 'POST',
             data: $("#contact-form").serialize(),
             success: function(data){
-            location.reload(true);
+             location.reload(true);
 
             }
      });
